@@ -54,17 +54,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.3.12",
+    "@angular/animations": "^18.2.11",
     "@angular/cdk": "^17.3.10",
-    "@angular/common": "^17.3.12",
-    "@angular/compiler": "^17.3.12",
-    "@angular/core": "^17.3.12",
-    "@angular/elements": "^17.3.12",
-    "@angular/forms": "^17.3.12",
+    "@angular/common": "^18.2.11",
+    "@angular/compiler": "^18.2.11",
+    "@angular/core": "^18.2.11",
+    "@angular/elements": "^18.2.11",
+    "@angular/forms": "^18.2.11",
     "@angular/material": "^17.3.10",
-    "@angular/platform-browser": "^17.3.12",
-    "@angular/platform-browser-dynamic": "^17.3.12",
-    "@angular/router": "^17.3.12",
+    "@angular/platform-browser": "^18.2.11",
+    "@angular/platform-browser-dynamic": "^18.2.11",
+    "@angular/router": "^18.2.11",
     "@senzing/rest-api-client-ng": "file:../libs/senzing-rest-api-client-ng-7.0.0-ng16-beta.0.tgz",
     "d3": "^7.8.4",
     "document-register-element": "^1.14.10",
@@ -76,10 +76,10 @@
     "zone.js": "~0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.11",
-    "@angular/cli": "~17.3.11",
-    "@angular/compiler-cli": "^17.3.12",
-    "@angular/language-service": "^17.3.12",
+    "@angular-devkit/build-angular": "^18.2.11",
+    "@angular/cli": "~18.2.11",
+    "@angular/compiler-cli": "^18.2.11",
+    "@angular/language-service": "^18.2.11",
     "@compodoc/compodoc": "^1.1.19",
     "@types/d3": "^7.4.0",
     "@types/jasmine": "~4.3.1",
@@ -94,7 +94,7 @@
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.0.0",
-    "ng-packagr": "^17.3.0",
+    "ng-packagr": "^18.2.1",
     "ngx-webstorage-service": "^5.0.0",
     "rimraf": "^6.0.1",
     "sass-burger": "^1.3.1",
