@@ -1,7 +1,7 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { ReactiveFormsModule, FormsModule } from '@angular/forms';
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { SenzingSdkModule, SzRestConfiguration  } from '@senzing/sdk-components-ng';
 import { AppComponent } from './app.component';
 import { SzExamplesHeader } from '../common/header.component';
@@ -28,25 +28,19 @@ export function SzRestConfigurationFactory() {
   return new SzRestConfiguration( (apiConfig ? apiConfig : undefined) );
 }
 
-@NgModule({
-  declarations: [
-    AppComponent,
-    SzExamplesHeader
-  ],
-  imports: [
-    BrowserModule,
-    HttpClientModule,
-    FormsModule,
-    ReactiveFormsModule,
-    SenzingSdkModule.forRoot( SzRestConfigurationFactory )
-  ],
-  providers: [
-    {
-      provide: HTTP_INTERCEPTORS,
-      useClass: MockTestDataInterceptor,
-      multi: true
-    }
-  ],
-  bootstrap: [AppComponent]
-})
+@NgModule({ declarations: [
+        AppComponent,
+        SzExamplesHeader
+    ],
+    bootstrap: [AppComponent], imports: [BrowserModule,
+        FormsModule,
+        ReactiveFormsModule,
+        SenzingSdkModule.forRoot(SzRestConfigurationFactory)], providers: [
+        {
+            provide: HTTP_INTERCEPTORS,
+            useClass: MockTestDataInterceptor,
+            multi: true
+        },
+        provideHttpClient(withInterceptorsFromDi())
+    ] })
 export class AppModule { }
