@@ -2,7 +2,7 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { ReactiveFormsModule, FormsModule } from '@angular/forms';
 import { CommonModule, TitleCasePipe } from '@angular/common';
-import { HttpClientModule, HttpClient } from '@angular/common/http';
+import { HttpClient, provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { OverlayModule } from '@angular/cdk/overlay';
 import { ApiModule as SenzingDataServiceModule } from '@senzing/rest-api-client-ng';
 import { SenzingSdkModule, SzRestConfiguration, SzPoweredByComponent, SzPrefsService, SzConfigurationService  } from '@senzing/sdk-components-ng';
@@ -30,24 +30,18 @@ export function SzRestConfigurationFactory() {
   return new SzRestConfiguration( (apiConfig ? apiConfig : undefined) );
 }
 
-@NgModule({
-  declarations: [
-    AppComponent
-  ],
-  imports: [
-    BrowserModule,
-    HttpClientModule,
-    FormsModule,
-    OverlayModule,
-    ReactiveFormsModule,
-    SenzingSdkModule.forRoot( SzRestConfigurationFactory ),
-    SenzingDataServiceModule.forRoot( SzRestConfigurationFactory )
-  ],
-  providers: [
-    SzConfigurationService,
-    SzDataMartService,
-    SzPrefsService
-  ],
-  bootstrap: [AppComponent]
-})
+@NgModule({ declarations: [
+        AppComponent
+    ],
+    bootstrap: [AppComponent], imports: [BrowserModule,
+        FormsModule,
+        OverlayModule,
+        ReactiveFormsModule,
+        SenzingSdkModule.forRoot(SzRestConfigurationFactory),
+        SenzingDataServiceModule.forRoot(SzRestConfigurationFactory)], providers: [
+        SzConfigurationService,
+        SzDataMartService,
+        SzPrefsService,
+        provideHttpClient(withInterceptorsFromDi())
+    ] })
 export class AppModule { }
